Planned Components Added
I went ahead and created some of the components I am planning on using in the application.

Header will hold the page's banner, website name, and a button to view the cart.

Cart will hold all the items the user chooses, in addition to the receipt of the items chosen with tax applied.

Body will hold a list of items that the user can choose to add to the cart. It will show individual Item components.

Item will hold the information regarding to the item which includes a picture, and the item's cost, with the button that lets the user add it into the cart.

The planned App state was also added into the App component.