Cleaned up Item Component CSS
The item components are now properly alligned with decent looking item sizes and font sizes to allow for better positioning and allignment.

Add To Cart buttons now have a cursor pointer when hovering over them.

Added a number next to the cart icon in the header to represent the amount of items that are currently in the cart.

Changed the price numbers for the items to strings, so that I can properly show decimal places on the item cards. These will be parsed to integers when time to calculate the total cart value.