Cart Dropdown Implementation
I decided to use the Materialize CSS Dropdown class. This required a small bit of extra CSS added into the App.css file to allow for the dropdown to not interfere with the other elements on the page.

The Cart component has been added to the drop down, and items added into the cart are properly being passed down via props.