Cart Items, Tax and Cost Functionality
The cart now properly calculates the tax and total sales cost of items in the cart. All imputs have been added as items and tested in the cart and are correctly showing the correct sales tax amount and total cart cost.

At this point, the entire functionality of the application has been finished. The rest of the time will be spent on cleaning up some of the CSS, misalignments with items and cleaning up the code to be a bit more readable.