Added new images, user feedback, CSS cleanup, README
I replaced some of the duplicate images with some new ones to separate the different imported perfumes and chocolates.

Fixed some of the spacing issues within the cart component so that the names and prices are spaced decently.

The Sales Tax and Total Cost text in the cart have been bolded and their text size increased a bit to be more readable for the user.

README has been added.